# packages/contracts/openapi/schema.yaml
openapi: 3.0.3
info:
  title: ECOnexão API
  version: 1.0.0
  description: API pública e administrativa versionada da ECOnexão.
paths:
...
  /api/v1/admin/auth/csrf:
    get:
      operationId: getAdminCsrfToken
      tags:
        - Admin auth
      security:
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
          description: ''
  /api/v1/admin/auth/login:
    post:
      operationId: loginAdmin
      tags:
        - Admin auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginError401'
          description: ''
  /api/v1/admin/auth/logout:
    post:
      operationId: logoutAdmin
      tags:
        - Admin auth
      security:
        - cookieAuth: []
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: ''
  /api/v1/admin/auth/session:
    get:
      operationId: getAdminSession
      tags:
        - Admin auth
      security:
        - cookieAuth: []
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: ''
  /api/v1/admin/dashboard/summary:
    get:
      operationId: api_v1_admin_dashboard_summary_retrieve
      description:
        Retorna contadores operacionais consolidados sem dados pessoais
        (PII) por região.
      parameters:
        - in: query
          name: region_slug
          schema:
            type: string
          description: Slug da região para filtrar os contadores operacionais.
      tags:
        - api
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
          description: ''
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '429':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '500':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
...
components:
  schemas:
    AdminIdentity:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: string
        region_slugs:
          type: array
          items:
            type: string
            pattern: ^[-a-zA-Z0-9_]+$
      required:
        - actions
        - id
        - region_slugs
        - roles
        - username
...
    CsrfResponse:
      type: object
      properties:
        csrf_token:
          type: string
      required:
        - csrf_token
...
    DashboardSummary:
      type: object
      properties:
        region_slug:
          type: string
        priority_reports_count:
          type: integer
        active_alerts_count:
          type: integer
        pending_revisions_count:
          type: integer
      required:
        - active_alerts_count
        - pending_revisions_count
        - priority_reports_count
        - region_slug
...
    LoginError401:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        field_errors:
          type: object
          additionalProperties: {}
        request_id:
          type: string
      required:
        - code
        - field_errors
        - message
        - request_id
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          writeOnly: true
          maxLength: 128
      required:
        - password
        - username
...
    SessionResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          allOf:
            - $ref: '#/components/schemas/AdminIdentity'
          nullable: true
      required:
        - authenticated
        - user
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: econexao_admin_sessionid